refactor: fix all remaining react-refresh lint warnings; make react-refresh violations an error
