+++ src/components/Base/index.ts
@@ -1,1 +1,3 @@
-export * from './Base'
+export { BaseComponent, type BaseComponentInterface, type CommonComponentInterface } from './Base'
+export { createCompoundContext } from './createCompoundContext'
+export { useBase } from './useBase'
