+++ src/components/Company/BankAccount/BankAccountForm/BankAccountForm.tsx
@@ -8,7 +8,8 @@ import { BankAccountFormSchema, Form } from './Form'
 import { Actions } from './Actions'
 import { BankAccountFormProvider } from './context'
 import type { BaseComponentInterface, CommonComponentInterface } from '@/components/Base/Base'
-import { BaseComponent, useBase } from '@/components/Base/Base'
+import { BaseComponent } from '@/components/Base/Base'
+import { useBase } from '@/components/Base/useBase'
 import { useI18n } from '@/i18n/I18n'
 import { Flex } from '@/components/Common/Flex/Flex'
 import { componentEvents } from '@/shared/constants'
