+++ src/components/Company/BankAccount/BankAccountList/BankAccountList.tsx
@@ -4,7 +4,8 @@ import { AccountView } from './AccountView'
 import { BankAccountProvider } from './context'
 import { Actions } from './Actions'
 import type { BaseComponentInterface, CommonComponentInterface } from '@/components/Base/Base'
-import { BaseComponent, useBase } from '@/components/Base/Base'
+import { BaseComponent } from '@/components/Base/Base'
+import { useBase } from '@/components/Base/useBase'
 import { useI18n } from '@/i18n/I18n'
 import { Flex } from '@/components/Common/Flex/Flex'
 import { componentEvents } from '@/shared/constants'
