+++ src/components/Company/BankAccount/BankAccountVerify/BankAccountVerify.tsx
@@ -8,7 +8,8 @@ import type { BankAccountVerifyInputs } from './Form'
 import { BankAccountVerifySchema, Form } from './Form'
 import { Actions } from './Actions'
 import type { BaseComponentInterface, CommonComponentInterface } from '@/components/Base/Base'
-import { BaseComponent, useBase } from '@/components/Base/Base'
+import { BaseComponent } from '@/components/Base/Base'
+import { useBase } from '@/components/Base/useBase'
 import { useI18n } from '@/i18n/I18n'
 import { Flex } from '@/components/Common/Flex/Flex'
 import { componentEvents } from '@/shared/constants'
