+++ src/components/Company/DocumentSignerFlow/DocumentList/DocumentList.tsx
@@ -8,7 +8,8 @@ import { Actions } from './Actions'
 import { DocumentListProvider } from './useDocumentList'
 import { useI18n } from '@/i18n'
 import type { CommonComponentInterface } from '@/components/Base/Base'
-import { useBase, BaseComponent, type BaseComponentInterface } from '@/components/Base/Base'
+import { BaseComponent, type BaseComponentInterface } from '@/components/Base/Base'
+import { useBase } from '@/components/Base/useBase'
 import { Flex } from '@/components/Common'
 import { companyEvents } from '@/shared/constants'
