+++ src/components/Company/Locations/LocationsList/LocationsList.tsx
@@ -6,7 +6,8 @@ import { Actions } from './Actions'
 import { LocationsListProvider } from './useLocationsList'
 import { useI18n } from '@/i18n'
 import type { CommonComponentInterface } from '@/components/Base/Base'
-import { BaseComponent, type BaseComponentInterface, useBase } from '@/components/Base/Base'
+import { BaseComponent, type BaseComponentInterface } from '@/components/Base/Base'
+import { useBase } from '@/components/Base/useBase'
 import { Flex } from '@/components/Common'
 import { companyEvents } from '@/shared/constants'
