+++ src/components/Company/Locations/locationsStateMachine.tsx
@@ -1,12 +1,10 @@
-import { transition, state, reduce } from 'robot3'
 import { LocationsList } from './LocationsList'
 import { LocationForm } from './LocationForm/LocationForm'
-import { companyEvents, componentEvents } from '@/shared/constants'
+import type { companyEvents } from '@/shared/constants'
 import { useFlowParams, type UseFlowParamsProps } from '@/components/Flow/hooks/useFlowParams'
 import type { FlowContextInterface } from '@/components/Flow/useFlow'
-import type { MachineEventType } from '@/types/Helpers'
 
-type EventPayloads = {
+export type EventPayloads = {
   [companyEvents.COMPANY_LOCATION_DONE]: undefined
   [companyEvents.COMPANY_LOCATION_EDIT]: { uuid: string }
   [companyEvents.COMPANY_LOCATION_CREATE]: undefined
@@ -34,63 +32,4 @@ export function LocationFormContextual() {
     requiredParams: ['companyId'],
   })
   return <LocationForm companyId={companyId} locationId={locationId} onEvent={onEvent} />
-}
-
-const cancelTransition = transition(
-  componentEvents.CANCEL,
-  'index',
-  reduce((ctx: LocationsContextInterface) => ({
-    ...ctx,
-    component: LocationsListContextual,
-    locationId: undefined,
-  })),
-)
-
-export const locationsStateMachine = {
-  index: state(
-    transition(
-      companyEvents.COMPANY_LOCATION_EDIT,
-      'locationEdit',
-      reduce(
-        (
-          ctx: LocationsContextInterface,
-          ev: MachineEventType<EventPayloads, typeof companyEvents.COMPANY_LOCATION_EDIT>,
-        ): LocationsContextInterface => ({
-          ...ctx,
-          component: LocationFormContextual,
-          locationId: ev.payload.uuid,
-        }),
-      ),
-    ),
-    transition(
-      companyEvents.COMPANY_LOCATION_CREATE,
-      'locationAdd',
-      reduce(
-        (ctx: LocationsContextInterface): LocationsContextInterface => ({
-          ...ctx,
-          component: LocationFormContextual,
-        }),
-      ),
-    ),
-  ),
-  locationAdd: state(
-    transition(
-      companyEvents.COMPANY_LOCATION_CREATED,
-      'index',
-      reduce((ctx: LocationsContextInterface) => ({ ...ctx, component: LocationsListContextual })),
-    ),
-    cancelTransition,
-  ),
-  locationEdit: state(
-    transition(
-      companyEvents.COMPANY_LOCATION_UPDATED,
-      'index',
-      reduce((ctx: LocationsContextInterface) => ({
-        ...ctx,
-        component: LocationsListContextual,
-        locationId: undefined,
-      })),
-    ),
-    cancelTransition,
-  ),
 }
