+++ src/components/Employee/PaymentMethod/Constants.ts
@@ -1,0 +1,9 @@
+export enum PAYMENT_METHODS {
+  check = 'Check',
+  directDeposit = 'Direct Deposit',
+}
+
+export enum SPLIT_BY {
+  percentage = 'Percentage',
+  amount = 'Amount',
+}
