+++ src/components/Flow/Flow.tsx
@@ -2,9 +2,9 @@ import { Suspense } from 'react'
 import { useMachine } from 'react-robot'
 import { type Machine } from 'robot3'
 import { Breadcrumb, Breadcrumbs, Link } from 'react-aria-components'
+import type { OnEventType } from '../Base/useBase'
 import type { FlowContextInterface } from './useFlow'
 import { FlowContext } from './useFlow'
-import type { OnEventType } from '@/components/Base'
 import { Loading } from '@/components/Common'
 import { type EventType } from '@/shared/constants'
